server_monitor: simplify get_name_info function
Previously, we checked for the presence of . in the hostname before we
checked for an osd address. This could fail for several reasons
including getting an IP address as a hostname or getting a more complex
hostname like (a.b.domain) where we truncated everything after the first
dot. The patch rewrites the function to deal with these problems.

Signed-off-by: Boris Ranto <branto@redhat.com>